requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
automaton==3.2.0
autopage==0.5.2
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cliff==4.9.1
cmd2==2.5.11
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1733502965406/work
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
datasets==3.5.0
debtcollector==3.0.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
dogpile.cache==1.3.4
et_xmlfile==2.0.0
eventlet==0.39.1
fasteners==0.19
filelock==3.18.0
filetype==1.2.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
futurist==3.1.0
greenlet==3.1.1
groq==0.24.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keystoneauth1==5.10.0
keystonemiddleware==10.9.0
kiwisolver==1.4.8
kombu==5.5.2
legacy-cgi==2.6.3
litellm==1.65.1
magnum==20.0.0
Mako==1.3.9
MarkupSafe==3.0.2
matplotlib==3.10.1
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
netaddr==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
openai==1.70.0
openpyxl==3.1.5
openstacksdk==4.4.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==3.2.0
oslo.cache==3.10.1
oslo.concurrency==7.1.0
oslo.config==9.7.1
oslo.context==5.7.1
oslo.db==17.2.1
oslo.i18n==6.5.1
oslo.log==7.1.0
oslo.messaging==16.1.0
oslo.metrics==0.11.0
oslo.middleware==6.3.1
oslo.policy==4.5.1
oslo.reports==3.5.1
oslo.serialization==5.7.0
oslo.service==4.1.1
oslo.upgradecheck==2.5.0
oslo.utils==8.2.0
oslo.versionedobjects==3.6.0
pandas==2.2.3
Paste==3.10.1
PasteDeploy==3.1.0
patsy==1.0.1
pbr==6.1.1
pecan==1.6.0
pillow==11.1.0
pipreqs==0.4.13
prettytable==3.16.0
prometheus_client==0.21.1
propcache==0.3.1
pyarrow==19.0.1
pycadf==4.0.1
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydot==3.0.4
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyperclip==1.9.0
python-barbicanclient==7.1.0
python-cinderclient==9.7.0
python-dotenv==1.1.0
python-glanceclient==4.8.0
python-heatclient==4.1.0
python-keystoneclient==5.6.0
python-neutronclient==11.4.0
python-novaclient==18.9.0
python-octaviaclient==3.10.0
python-openstackclient==7.4.0
python-swiftclient==4.7.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
repoze.lru==0.7
requests==2.32.3
requestsexceptions==1.4.0
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.24.0
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
simplegeneric==0.8.1
sniffio==1.3.1
SQLAlchemy==2.0.40
statsd==4.0.1
statsmodels==0.14.4
stevedore==5.4.1
taskflow==5.12.0
tenacity==9.1.2
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
trimesh==4.6.6
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
warlock==2.0.1
WebOb==1.8.9
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
WSME==0.12.1
xxhash==3.5.0
yappi==1.6.10
yarg==0.1.10
yarl==1.18.3